@@ -10,23 +10,23 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.1.0",
+    "@abacritt/angularx-social-login": "^1.3.2",
+    "@angular/animations": "^15.2.0",
     "@angular/cdk": "^13.1.3",
-    "@angular/common": "~13.1.0",
-    "@angular/compiler": "~13.1.0",
-    "@angular/core": "~13.1.0",
-    "@angular/forms": "~13.1.0",
+    "@angular/common": "^15.2.0",
+    "@angular/compiler": "^15.2.0",
+    "@angular/core": "^15.2.0",
+    "@angular/forms": "^15.2.0",
     "@angular/material": "^13.1.3",
     "@angular/material-moment-adapter": "^13.3.2",
-    "@angular/platform-browser": "~13.1.0",
-    "@angular/platform-browser-dynamic": "~13.1.0",
-    "@angular/router": "~13.1.0",
+    "@angular/platform-browser": "^15.2.0",
+    "@angular/platform-browser-dynamic": "^15.2.0",
+    "@angular/router": "^15.2.0",
     "angular-cli-ghpages": "^1.0.0",
-    "angularx-social-login": "^4.0.1",
     "docxtemplater": "^3.28.6",
     "file-saver": "^2.0.5",
     "moment": "^2.29.2",
-    "ngx-material-file-input": "^3.0.0",
+    "ngx-material-file-input": "^4.0.0",
     "pizzip": "^3.1.1",
     "rxjs": "~7.4.0",
     "sweetalert2": "^11.3.6",
@@ -34,9 +34,9 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.1.3",
-    "@angular/cli": "~13.1.3",
-    "@angular/compiler-cli": "~13.1.0",
+    "@angular-devkit/build-angular": "^15.2.0",
+    "@angular/cli": "^15.2.0",
+    "@angular/compiler-cli": "^15.2.0",
     "@types/jasmine": "~3.10.0",
     "@types/node": "^12.11.1",
     "jasmine-core": "~3.10.0",
@@ -45,6 +45,6 @@
     "karma-coverage": "~2.1.0",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "~1.7.0",
-    "typescript": "~4.5.2"
+    "typescript": "~4.8.4"
   }
 }
