@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import {FormBuilder, FormControl, FormGroup, Validators} from "@angular/forms";
+import {UntypedFormBuilder, UntypedFormControl, UntypedFormGroup, Validators} from "@angular/forms";
 import {ActivatedRoute, Router} from "@angular/router";
 import {FechaService} from "../../../services/fecha.service";
 import {ProyectoService} from "../../../services/proyecto.service";
@@ -50,25 +50,25 @@ function loadFile(url:any, callback:any) {
 export class DelegaciondealumnosComponent implements OnInit {
 
   isLinear = true;
-  firstFormGroup?: FormGroup;
-  secondFormGroup?: FormGroup;
-  thirtdFormGroup?: FormGroup;
-  fourFormGroup?: FormGroup;
+  firstFormGroup?: UntypedFormGroup;
+  secondFormGroup?: UntypedFormGroup;
+  thirtdFormGroup?: UntypedFormGroup;
+  fourFormGroup?: UntypedFormGroup;
 
   issloading=true;
   isexist?:boolean
 
-  myControlproyecto = new FormControl();
+  myControlproyecto = new UntypedFormControl();
   filteredOptionsProyecto?: Observable<Proyectos[]>;
   proyectos:Proyectos[]=[];
   proyectoselect:Proyectos=new Proyectos();
 
-  myControlanexo1 = new FormControl();
+  myControlanexo1 = new UntypedFormControl();
   filteredOptionsanexo1?: Observable<Anexo1[]>;
   docentesAnexo1:Anexo1[]=[];
   docenteselect:Anexo1 = new Anexo1();
 
-  myControlAnexe4 = new FormControl();
+  myControlAnexe4 = new UntypedFormControl();
   filteredOptionsAnexe4?: Observable<Anexo4[]>;
   alumnosAnexe4:Anexo4[]=[];
   alumnoselect:Anexo4[]=[];
@@ -82,7 +82,7 @@ export class DelegaciondealumnosComponent implements OnInit {
               private activatedRoute: ActivatedRoute,
               private proyectoService:ProyectoService,
               private responsablepppService:ResponsablepppService,
-              private _formBuilder: FormBuilder,
+              private _formBuilder: UntypedFormBuilder,
               private entidadbeneficiarioService:EntidadbeneficiarioService,
               private _adapter: DateAdapter<any>,
               private anexo2Service:Anexo2Service,
