@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import {Anexo9} from "../../../models/anexo9";
-import {FormBuilder, FormControl} from "@angular/forms";
+import {UntypedFormBuilder, UntypedFormControl} from "@angular/forms";
 import {map, Observable, startWith} from "rxjs";
 import {InformeSeguimiento} from "../../../models/seguimiento";
 import {ActivatedRoute, Router} from "@angular/router";
@@ -33,13 +33,13 @@ export class ListainformeseguimientoComponent implements OnInit {
 cedulaa?:String;
   nombress?:String;
   informeS:InformeSeguimiento[]=[];
-  myControl = new FormControl();
+  myControl = new UntypedFormControl();
   filteredOptions?: Observable<InformeSeguimiento[]>;
 
   constructor(private router:Router,
               private activatedRoute: ActivatedRoute,
               private proyectoService: ProyectoService,
-              private _formBuilder: FormBuilder,
+              private _formBuilder: UntypedFormBuilder,
               private _adapter: DateAdapter<any>,
               private informeService:InformeSeguimientoService) { }
 
