@@ -3,7 +3,7 @@ import {ActivatedRoute, Router} from "@angular/router";
 import {FechaService} from "../../../services/fecha.service";
 import {ProyectoService} from "../../../services/proyecto.service";
 import {ResponsablepppService} from "../../../services/responsableppp.service";
-import {FormBuilder, FormControl} from "@angular/forms";
+import {UntypedFormBuilder, UntypedFormControl} from "@angular/forms";
 import {EntidadbeneficiarioService} from "../../../services/entidadbeneficiario.service";
 import {DateAdapter} from "@angular/material/core";
 import {Anexo2Service} from "../../../services/anexo2.service";
@@ -68,9 +68,9 @@ export class Verpostulaciones1Component implements OnInit {
   // @ts-ignore
   @ViewChild(MatSort) sort: MatSort;
 
-  myControla = new FormControl();
+  myControla = new UntypedFormControl();
   filteredOptionsa?: Observable<Anexo3[]>;
-  myControlr = new FormControl();
+  myControlr = new UntypedFormControl();
   filteredOptionsr?: Observable<Anexo3[]>;
 
   constructor(private router: Router,
@@ -78,7 +78,7 @@ export class Verpostulaciones1Component implements OnInit {
               private activatedRoute: ActivatedRoute,
               private proyectoService:ProyectoService,
               private responsablepppService:ResponsablepppService,
-              private _formBuilder: FormBuilder,
+              private _formBuilder: UntypedFormBuilder,
               private entidadbeneficiarioService:EntidadbeneficiarioService,
               private _adapter: DateAdapter<any>,
               private anexo2Service:Anexo2Service,
