@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import {FormBuilder} from "@angular/forms";
+import {UntypedFormBuilder} from "@angular/forms";
 import {ActivatedRoute, Router} from "@angular/router";
 import {FechaService} from "../../../services/fecha.service";
 import {ProyectoService} from "../../../services/proyecto.service";
@@ -25,7 +25,7 @@ export class VerseleccionestudiabesComponent implements OnInit {
 
   issloading=true;
   informeAceotacion:Informeaceptacion[] = [];
-  constructor(private _formBuilder: FormBuilder,
+  constructor(private _formBuilder: UntypedFormBuilder,
               private router: Router,
               private fechaService:FechaService,
               private activatedRoute: ActivatedRoute,
